fix : emotion api

+++ src/main/java/com/bestoctopus/dearme/service/DailyEmoService.java
@@ -2,7 +2,10 @@ package com.bestoctopus.dearme.service;
 
 import com.bestoctopus.dearme.domain.DailyEmo;
 import com.bestoctopus.dearme.domain.Emotion;
+import com.bestoctopus.dearme.domain.User;
+import com.bestoctopus.dearme.dto.DailyEmoDto;
 import com.bestoctopus.dearme.repository.DailyEmoRepository;
+import com.bestoctopus.dearme.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,6 +18,8 @@ import java.util.*;
 @RequiredArgsConstructor
 public class DailyEmoService {
     private final DailyEmoRepository dailyEmoRepository;
+
+    private final UserRepository userRepository;
 
     public List<DailyEmo> getAllEmotionList(LocalDate startDate, LocalDate endDate) {
         return dailyEmoRepository.findBydateBetween(startDate, endDate);
@@ -40,32 +45,33 @@ public class DailyEmoService {
         return new ArrayList<>(emotionCountMap.entrySet());
     }
 
-    public DailyEmo postDailyEmo(DailyEmo dailyEmo) {
+    public DailyEmo postDailyEmo(DailyEmoDto dailyEmoDto, String user_id) {
+        User user = userRepository.findById(user_id).orElseThrow();
+
+        DailyEmo dailyEmo = dailyEmoDto.toEntity(user);
+
         return dailyEmoRepository.save(dailyEmo);
     }
 
-    public boolean putDailyEmo(Long emotionId, Emotion emotion) {
+    /*
+    public boolean deleteDailyEmo(Long emotionId, String user_id) {
+        User user = userRepository.findById(user_id).orElseThrow();
+
         Optional<DailyEmo> optionalDailyEmo = dailyEmoRepository.findById(emotionId);
 
         if (optionalDailyEmo.isPresent()) {
             DailyEmo dailyEmo = optionalDailyEmo.get();
-            dailyEmo.update(emotion);
-            dailyEmoRepository.save(dailyEmo);
-            return true;
+            if (dailyEmo.getUser().equals(user)) {
+                dailyEmoRepository.deleteById(emotionId);
+                return true;
+            }
+            else {
+                return false;
+            }
         } else {
             return false;
         }
     }
-
-    public boolean deleteDailyEmo(Long emotionId) {
-        Optional<DailyEmo> optionalDailyEmo = dailyEmoRepository.findById(emotionId);
-
-        if (optionalDailyEmo.isPresent()) {
-            dailyEmoRepository.deleteById(emotionId);
-            return true;
-        } else {
-            return false;
-        }
-    }
+     */
 
 }
